Moved institution admins controller to institutions namespace
